+++ warden/x/warden/keeper/msg_server_new_key_request.go
@@ -23,22 +23,25 @@ func (k msgServer) NewKeyRequest(ctx context.Context, msg *types.MsgNewKeyReques
 		return nil, err
 	}
 
+	if err := keychain.EnsureSufficientKeyFees(msg.MaxKeychainFees); err != nil {
+		return nil, err
+	}
+
 	if err := k.deductKeychainFees(
 		ctx,
-		msg.MaxKeychainFees,
-		keychain.AccAddress(),
-		keychain.Fees.KeyReq,
-		sdk.MustAccAddressFromBech32(creator)); err != nil {
+		sdk.MustAccAddressFromBech32(creator),
+		keychain.Fees.KeyReq); err != nil {
 		return nil, err
 	}
 
 	req := &types.KeyRequest{
-		Creator:    creator,
-		SpaceId:    msg.SpaceId,
-		KeychainId: msg.KeychainId,
-		KeyType:    msg.KeyType,
-		Status:     types.KeyRequestStatus_KEY_REQUEST_STATUS_PENDING,
-		RuleId:     msg.RuleId,
+		Creator:              creator,
+		SpaceId:              msg.SpaceId,
+		KeychainId:           msg.KeychainId,
+		KeyType:              msg.KeyType,
+		Status:               types.KeyRequestStatus_KEY_REQUEST_STATUS_PENDING,
+		RuleId:               msg.RuleId,
+		DeductedKeychainFees: keychain.Fees.KeyReq,
 	}
 
 	id, err := k.keyRequests.Append(ctx, req)