+++ warden/x/warden/types/v1beta3/keychain.go
@@ -98,3 +98,19 @@ func NewKeybaseId(value string) (*KeybaseId, error) {
 
 	return &KeybaseId{Value: value}, nil
 }
+
+func (k *Keychain) EnsureSufficientKeyFees(fees sdk.Coins) error {
+	return ensureSufficientFees(k.Fees.KeyReq, fees)
+}
+
+func (k *Keychain) EnsureSufficientSignFees(fees sdk.Coins) error {
+	return ensureSufficientFees(k.Fees.SigReq, fees)
+}
+
+func ensureSufficientFees(wantedFees sdk.Coins, maxFees sdk.Coins) error {
+	if wantedFees.IsAllLTE(maxFees) {
+		return nil
+	}
+
+	return fmt.Errorf("keychain fees are not sufficient: wanted %s", wantedFees)
+}